refactor(models): ParamsB int → float64 (preserve sub-1B precision)
int storage truncated qwen3-0.6b to "0 B" in `list` and lost fidelity
for any sub-1B or fractional model (e.g., gemma-4-E4B at 7.5B reading
as 7). Migrate ParamsB to float64 across Metadata, Model, callers, and
test fixtures. JSON deserialization handles old int values transparently.
QuantSizeTable stays keyed by int; selector converts via math.Round at
lookup. Display uses %g for terse output.

+++ internal/models/whitelist.go
@@ -11,24 +11,24 @@ type Model struct {
 	ID      string // canonical llamactl id, e.g. "qwen2.5-7b-instruct"
 	HFRepo  string // HuggingFace repo, e.g. "Qwen/Qwen2.5-7B-Instruct-GGUF"
 	Arch    Arch
-	ParamsB int // parameter count in billions, must have a row in QuantSizeTable
-	MaxCtx  int // maximum context tokens supported by the model family
+	ParamsB float64 // parameter count in billions, must have a row in QuantSizeTable
+	MaxCtx  int     // maximum context tokens supported by the model family
 }
 
 // PreferredIDs is the curated table of short-ids with pre-tuned ParamsB/Arch/MaxCtx.
 // Adding entries is a code change; `add` is not gated by this table — any HF GGUF repo
 // path is accepted with explicit --quant (see PRD §4).
 var PreferredIDs = map[string]Model{
-	"qwen2.5-3b-instruct":  {ID: "qwen2.5-3b-instruct", HFRepo: "Qwen/Qwen2.5-3B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 3, MaxCtx: 32768},
-	"qwen2.5-7b-instruct":  {ID: "qwen2.5-7b-instruct", HFRepo: "Qwen/Qwen2.5-7B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 7, MaxCtx: 32768},
-	"qwen2.5-14b-instruct": {ID: "qwen2.5-14b-instruct", HFRepo: "Qwen/Qwen2.5-14B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 14, MaxCtx: 32768},
-	"qwen2.5-coder-3b":     {ID: "qwen2.5-coder-3b", HFRepo: "Qwen/Qwen2.5-Coder-3B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 3, MaxCtx: 32768},
-	"qwen2.5-coder-7b":     {ID: "qwen2.5-coder-7b", HFRepo: "Qwen/Qwen2.5-Coder-7B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 7, MaxCtx: 32768},
-	"qwen2.5-coder-14b":    {ID: "qwen2.5-coder-14b", HFRepo: "Qwen/Qwen2.5-Coder-14B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 14, MaxCtx: 32768},
-	"llama3.1-8b":          {ID: "llama3.1-8b", HFRepo: "bartowski/Meta-Llama-3.1-8B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 8, MaxCtx: 131072},
-	"llama3.2-3b":          {ID: "llama3.2-3b", HFRepo: "bartowski/Llama-3.2-3B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 3, MaxCtx: 131072},
-	"llama3.3-70b":         {ID: "llama3.3-70b", HFRepo: "bartowski/Llama-3.3-70B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 70, MaxCtx: 131072},
-	"mistral-7b-v0.3":      {ID: "mistral-7b-v0.3", HFRepo: "bartowski/Mistral-7B-Instruct-v0.3-GGUF", Arch: ArchMistral, ParamsB: 7, MaxCtx: 32768},
+	"qwen2.5-3b-instruct":  {ID: "qwen2.5-3b-instruct", HFRepo: "Qwen/Qwen2.5-3B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 3.0, MaxCtx: 32768},
+	"qwen2.5-7b-instruct":  {ID: "qwen2.5-7b-instruct", HFRepo: "Qwen/Qwen2.5-7B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 7.0, MaxCtx: 32768},
+	"qwen2.5-14b-instruct": {ID: "qwen2.5-14b-instruct", HFRepo: "Qwen/Qwen2.5-14B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 14.0, MaxCtx: 32768},
+	"qwen2.5-coder-3b":     {ID: "qwen2.5-coder-3b", HFRepo: "Qwen/Qwen2.5-Coder-3B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 3.0, MaxCtx: 32768},
+	"qwen2.5-coder-7b":     {ID: "qwen2.5-coder-7b", HFRepo: "Qwen/Qwen2.5-Coder-7B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 7.0, MaxCtx: 32768},
+	"qwen2.5-coder-14b":    {ID: "qwen2.5-coder-14b", HFRepo: "Qwen/Qwen2.5-Coder-14B-Instruct-GGUF", Arch: ArchQwen25, ParamsB: 14.0, MaxCtx: 32768},
+	"llama3.1-8b":          {ID: "llama3.1-8b", HFRepo: "bartowski/Meta-Llama-3.1-8B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 8.0, MaxCtx: 131072},
+	"llama3.2-3b":          {ID: "llama3.2-3b", HFRepo: "bartowski/Llama-3.2-3B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 3.0, MaxCtx: 131072},
+	"llama3.3-70b":         {ID: "llama3.3-70b", HFRepo: "bartowski/Llama-3.3-70B-Instruct-GGUF", Arch: ArchLlama3, ParamsB: 70.0, MaxCtx: 131072},
+	"mistral-7b-v0.3":      {ID: "mistral-7b-v0.3", HFRepo: "bartowski/Mistral-7B-Instruct-v0.3-GGUF", Arch: ArchMistral, ParamsB: 7.0, MaxCtx: 32768},
 }
 
 // LookupOrSuggest returns the PreferredIDs entry for id, or an error listing
